requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
pydantic>=2.9,<3.0
pydantic-settings>=2.6,<3.0
httpx>=0.28,<1.0
pypdfium2>=4.30,<5.0
PyMuPDF>=1.24,<2.0
Pillow>=11.0,<12.0
numpy>=1.26,<3.0
python-multipart>=0.0.20,<1.0
python-dateutil>=2.9,<3.0
python-slugify>=8.0,<9.0
orjson>=3.10,<4.0
alembic>=1.14,<2.0
huggingface_hub>=1.5,<2.0
transformers>=5.2,<6.0
accelerate>=1.11,<2.0
# Chandra OCR 2 is on the upstream GitHub repo before a matching PyPI 0.2.x release.
# Pin the exact commit that was verified in Docker build for reproducibility.
# Torch/Torchvision are installed separately in Docker from the CUDA wheel index.
chandra-ocr @ git+https://github.com/datalab-to/chandra.git@01f86ebd3a9eef07aa69734008a86541f206456e
pytest>=9.0,<10.0
